refactor: isolate browser PAT crypto
