+++ my-app/src/lib/mcp/tokens.test.ts
@@ -2,9 +2,6 @@
 import { describe, expect, test } from "vitest";
 import { exportPKCS8, generateKeyPair, jwtVerify, createLocalJWKSet } from "jose";
 import {
-  sha256Hex,
-  randomToken,
-  randomHex,
   mintAccessToken,
   mintPatBridgeToken,
   getPublicJwks,
@@ -17,26 +14,6 @@ async function testPem(): Promise<string> {
   const { privateKey } = await generateKeyPair("RS256", { extractable: true });
   return await exportPKCS8(privateKey);
 }
-
-describe("sha256Hex", () => {
-  test("hashes to the known vector for 'abc'", async () => {
-    expect(await sha256Hex("abc")).toBe(
-      "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad",
-    );
-  });
-});
-
-describe("randomToken / randomHex", () => {
-  test("randomToken is base64url and unique", () => {
-    const a = randomToken();
-    expect(a).toMatch(/^[A-Za-z0-9_-]{43}$/); // 32 bytes → 43 base64url chars
-    expect(randomToken()).not.toBe(a);
-  });
-
-  test("randomHex is lowercase hex of requested length", () => {
-    expect(randomHex(16)).toMatch(/^[0-9a-f]{32}$/);
-  });
-});
 
 describe("mintAccessToken", () => {
   test("mints an RS256 JWT verifiable via the derived JWKS with expected claims", async () => {
