+++ my-app/src/lib/mcp/tokens.ts
@@ -1,41 +1,13 @@
 // src/lib/mcp/tokens.ts
-// Crypto + JWT helpers for the MCP OAuth server. Isomorphic (Web Crypto only):
-// runs in Next.js route handlers, server actions, edge-runtime tests, and the
-// browser (PAT generation in the settings UI). Convex functions never import
-// this — they only ever receive pre-computed hashes.
+// Server-only JWT helpers for the MCP OAuth server. Convex functions never
+// import this; they only receive minted credentials.
+import "server-only";
 import { SignJWT, importPKCS8, exportJWK } from "jose";
 
 export const MCP_JWT_AUDIENCE = "fragrances-mcp";
 export const MCP_JWT_KID = "mcp-1";
 export const ACCESS_TOKEN_TTL_SECONDS = 900; // 15 min
 export const PAT_BRIDGE_TTL_SECONDS = 300; // 5 min
-export const PAT_PREFIX = "fgt_";
-
-export async function sha256Hex(input: string): Promise<string> {
-  const digest = await crypto.subtle.digest("SHA-256", new TextEncoder().encode(input));
-  return Array.from(new Uint8Array(digest))
-    .map((b) => b.toString(16).padStart(2, "0"))
-    .join("");
-}
-
-function randomBytes(n: number): Uint8Array {
-  const bytes = new Uint8Array(n);
-  crypto.getRandomValues(bytes);
-  return bytes;
-}
-
-/** Opaque secret (refresh tokens, auth codes, PAT bodies): base64url, no padding. */
-export function randomToken(bytes = 32): string {
-  const bin = String.fromCharCode(...randomBytes(bytes));
-  return btoa(bin).replaceAll("+", "-").replaceAll("/", "_").replaceAll("=", "");
-}
-
-/** Public identifiers (client_id): lowercase hex. */
-export function randomHex(bytes = 16): string {
-  return Array.from(randomBytes(bytes))
-    .map((b) => b.toString(16).padStart(2, "0"))
-    .join("");
-}
 
 function requireEnv(value: string | undefined, name: string): string {
   if (!value) throw new Error(`Missing required env var ${name}.`);
